refactor(synthesize): EvidenceCollectorService — one contract for all non-fact prompt sections
The orchestrator was accreting one @Optional lane service per evidence
class (episode quotes, provenance excerpts, segments, mention scan,
insights, instruction probe — 7 constructor deps by V9). Every typed
prompt section beyond the fact lines now comes from ONE collector
behind one contract:

- EvidenceCollectorService.collect() → { transcriptLines,
  insightLines, instructions, timelineEvidence } — the three collect*
  method bodies moved verbatim (same gating via the pure
  evidence-gates module, same degrade-to-empty contracts, same
  intra-collect concurrency), so adding an evidence lane now touches
  the collector only.
- timelineEvidence is computed exactly once in the contract — the
  generator and verifier MENTION RECORD framings can no longer
  disagree by construction.
- The instruction probe now starts AFTER the empty/no-results and
  coverage-abstention exits instead of being fired early and
  deliberately abandoned on those paths — the pre-V9 wasted search is
  gone along with the dangling-promise idiom.
- SynthesizeService: 7 ctor deps → 4; positionally-constructed unit
  tests unchanged.

Behavior-neutral: full suite green (233/1933), tsc + lint clean. dist
still deliberately not rebuilt until the V9 measurement queue drains.
